# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a Zombies server.",
    "version": "1.0.0",
    "title": "Swagger Zombies",
    "contact": {
      "email": "tbuslowski@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/api/v1",
  "host": "localhost:5000",
  "schemes": [
    "http",
    "https"
  ],
  "tags": [
    {
      "name": "zombies",
      "description": "Everything about zombies",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://swagger.io"
      }
    },
    {
      "name": "items",
      "description": "Everything about available items",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://swagger.io"
      }
    },
    {
      "name": "rates",
      "description": "Everything about actual rates",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://swagger.io"
      }
    }
  ],
  "paths": {
    "/zombies": {
      "get": {
        "tags": [
          "zombies"
        ],
        "summary": "List of zombies",
        "description": "",
        "operationId": "listOfZombies",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "zombies"
        ],
        "summary": "Add a new zombie to the db",
        "description": "",
        "operationId": "addZombie",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Zombie object that needs to be added to the db",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ZombiePost"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Zombie"
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/zombies/{zombieId}": {
      "get": {
        "tags": [
          "zombies"
        ],
        "summary": "Find zombie by ID",
        "description": "Returns a single zombie",
        "operationId": "getZombieById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "zombieId",
            "in": "path",
            "description": "ID of zombie to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Zombie"
            }
          },
          "404": {
            "description": "Zombie not found"
          }
        }
      },
      "put": {
        "tags": [
          "zombies"
        ],
        "summary": "Update an existing zombie",
        "description": "",
        "operationId": "updateZombie",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Zombie object that needs to be changed in the db",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ZombiePut"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Zombie"
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Zombie not found"
          }
        }
      },
      "delete": {
        "tags": [
          "zombies"
        ],
        "summary": "Deletes a zombie",
        "description": "",
        "operationId": "deleteZombie",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "zombieId",
            "in": "path",
            "description": "Zombie id to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Zombie deleted"
          },
          "404": {
            "description": "Zombie not found"
          }
        }
      }
    },
    "/items": {
      "get": {
        "tags": [
          "items"
        ],
        "summary": "List of available items",
        "description": "",
        "operationId": "listOfItems",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/items/{zombieId}": {
      "get": {
        "tags": [
          "items"
        ],
        "summary": "List item for zombieId",
        "description": "",
        "operationId": "listOfZombieItems",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "zombieId",
            "in": "path",
            "description": "ID of zombie to return his items",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Item"
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "post": {
        "tags": [
          "items"
        ],
        "parameters": [
          {
            "name": "zombieId",
            "in": "path",
            "description": "ID of zombie to return his items",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Items array",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ItemId"
              }
            }
          }
        ],
        "summary": "Set new items for zombie",
        "description": "",
        "operationId": "createZombieItems",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Item"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "items"
        ],
        "summary": "Update items for zombie",
        "description": "",
        "operationId": "updateZombieItems",
        "parameters": [
          {
            "name": "zombieId",
            "in": "path",
            "description": "ID of zombie to return his items",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Items array",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ItemId"
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Item"
              }
            }
          }
        }
      }
    },
    "/items/{zombieId}/{itemId}": {
      "delete": {
        "tags": [
          "items"
        ],
        "summary": "Delete item from zombie items",
        "description": "",
        "operationId": "deleteItemFromZombie",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "zombieId",
            "in": "path",
            "description": "ID of zombie to remove his item",
            "required": true,
            "type": "string"
          },
          {
            "name": "itemId",
            "in": "path",
            "description": "ID of item to remove",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Item"
              }
            }
          }
        }
      }
    },
    "/rates": {
      "get": {
        "tags": [
          "rates"
        ],
        "summary": "List of actual rates",
        "description": "",
        "operationId": "listOfRates",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Rate": {
      "type": "object",
      "properties": {
        "pln": {
          "type": "number",
          "format": "float"
        },
        "usd": {
          "type": "number",
          "format": "float"
        },
        "eur": {
          "type": "number",
          "format": "float"
        }
      },
      "xml": {
        "name": "Rate"
      }
    },
    "Item": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "integer"
        }
      },
      "xml": {
        "name": "Item"
      }
    },
    "ItemId": {
      "type": "number",
      "xml": {
        "name": "ItemId"
      }
    },
    "Zombie": {
      "type": "object",
      "required": [
        "name",
        "items"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string",
          "minLength": 2
        },
        "items": {
          "type": "array",
          "maxItems": 5,
          "items": {
            "$ref": "#/definitions/Item"
          }
        },
        "itemsTotalValue": {
          "type": "object",
          "$ref": "#/definitions/Rate"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "xml": {
        "name": "Zombie"
      }
    },
    "ZombiePost": {
      "type": "object",
      "required": [
        "name",
        "items"
      ],
      "properties": {
        "name": {
          "type": "string",
          "minLength": 2
        },
        "items": {
          "type": "array",
          "maxItems": 5,
          "items": {
            "$ref": "#/definitions/ItemId"
          }
        }
      },
      "xml": {
        "name": "ZombiePost"
      }
    },
    "ZombiePut": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "name": {
          "type": "string",
          "minLength": 2
        }
      },
      "xml": {
        "name": "ZombiePut"
      }
    },
    "ApiResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}